requirements.txt
AlarmEZ

Flask-SQLAlchemy==1.0